tests/wpt/tests/css/css-transforms/parsing/webkit-perspective-invalid.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Compatibility: parsing -webkit-perspective with invalid values</title>
<link rel="help" href="https://compat.spec.whatwg.org/#propdef--webkit-perspective">
<link rel="help" href="https://github.com/whatwg/compat/issues/100">
<meta name="assert" content="-webkit-perspective also supports '<number [0,∞]>' in addition to 'none | <length [0,∞]>', but not 'calc([ <number [0,∞]> ])'">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("-webkit-perspective", "calc(1000)");
test_invalid_value("-webkit-perspective", "calc(25)");
</script>
</body>
</html>
